refactor(decomp): extract cdc sinks → proximadb-cdc (TD-DECOMP-9)
Moves src/cdc/sinks/{mod,kafka,webhook,traits} (44 tests) into the cdc
crate (83→127 tests). Whole subdir, verbatim; rewrite crate::cdc::{event,
error}→crate::{event,error}. Added deps: async-trait, tracing, reqwest
(webhook HTTP, inline 0.11/json). Added cdc-kafka feature (cfg gate for
the rdkafka-backed Kafka impl, mirrors root). Root re-exports sinks.

Verified: cdc crate 127 tests pass; root check + server build + cdc-clippy
+ fmt + check_workspace_boundaries + check-layering all green.
