+++ crates/integrations/proximadb-cdc/src/lib.rs
@@ -15,4 +15,5 @@
 pub mod error;
 pub mod event;
 pub mod metrics;
+pub mod sinks;
 pub mod transform;
