+++ docs/10-quality/td/TD-DECOMP-9-extract-cdc-sinks.adoc
@@ -1,0 +1,41 @@
+// Copyright (C) 2025 ProximaDB
+// SPDX-License-Identifier: Apache-2.0
+= TD-DECOMP-9: Extract cdc sinks → proximadb-cdc (god-crate decomposition)
+
+[cols="1,3", options="header"]
+|===
+| Field | Value
+| Status | Landing (delivered by this slice)
+| Severity | Low — build/dev-loop only; zero runtime, API, or on-disk-format impact (behavior-neutral re-export)
+| Component | `crates/integrations/proximadb-cdc` (`sinks`); root `src/cdc/mod.rs` (re-export shim)
+| Relates to | `roadmap/techdebt/P2_GOD_CRATE_TEST_DECOMPOSITION.md`; [[TD-DECOMP-1]] (recipe); [[TD-DECOMP-7]]/[[TD-DECOMP-8]] (cdc crate seed + transform/error)
+|===
+
+== Why
+
+Continues the cdc vein. The sinks subgraph (`mod.rs` + `kafka` + `webhook` + `traits`, **44
+tests**) is the next cohesive unit — its only non-self deps are `event` + `error` (both now
+in the cdc crate via TD-DECOMP-7/8). Deps: `serde` + `tracing` + `async-trait` + `reqwest`
+(webhook HTTP). The cdc crate goes 83→127 tests.
+
+== What moved (behavior-neutral)
+
+* `src/cdc/sinks/` (`mod.rs` 6t + `kafka.rs` 14t + `webhook.rs` 16t + `traits.rs` 8t) →
+  `crates/integrations/proximadb-cdc/src/sinks/` (verbatim subdir move).
+* Rewrite: `crate::cdc::{event,error}` → `crate::{event,error}` (now siblings).
+* `proximadb-cdc/Cargo.toml`: + `async-trait`, `tracing` (workspace), `reqwest` (inline
+  `{ version = "0.11", features = ["json"] }` — not a workspace dep).
+* `proximadb-cdc/src/lib.rs`: `pub mod sinks;`.
+* Root `src/cdc/mod.rs`: `pub mod sinks;` → `pub use proximadb_cdc::sinks;`.
+
+== Verification
+
+* `cargo check -p proximadb-cdc` green; `nextest` — 127 tests pass.
+* `cargo check -p proximadb` green (root compiles through re-export).
+* `check_workspace_boundaries.py` + `check-layering.sh` green (integration-tier-legal).
+* `cargo build -p proximadb-server` + `clippy -D warnings` + `fmt` + `work-commit-check` green.
+
+== Follow-ups (cdc vein)
+
+* `outbound/{dedup,router,config,position,subscriber,exactly_once}` + `connectors/*` + `coordinator`.
+* **Running total once merged: ~252 tests out of root lib binary** (208 + 44).
